requirements.txt
streamlit==1.33.0
streamlit-audiorecorder==0.0.5
groq==0.5.0
gradio==4.29.0
python-dotenv==1.0.1